Fix "last step wins" error (#132)
* Add a test case that has a pinned action after an un-pinned one

Whether the job fails or not was being overwritten by the last step in
the job. If the last step is pinned, this would cause the check to
pass, even if an earlier step was unpinned. Now we have a test case that
exercises this functionality.

* Fail the job if any steps have errors

Previously this logic only failed the job if the last step had an error.

* Run npm prepare

---------

Co-authored-by: Zennon Gosalvez <1798166+zgosalvez@users.noreply.github.com>